backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.0
openai==1.50.0
numpy==1.26.4
faiss-cpu==1.8.0
zeroentropy>=0.1.0
